Tighten up semantics of the interface.
git-svn-id: https://europa-pso.googlecode.com/svn/PLASMA/trunk@3409 23e1bc29-d954-0410-9766-47de34158ba0

+++ ConstraintEngine/test/domain-tests.cc
@@ -63,7 +63,7 @@ namespace EUROPA {
       IntervalDomain realDomain(10.2,20.4);
       assertFalse(realDomain.isEmpty());
       assertFalse(realDomain.isFinite());
-      assert(realDomain.getSize() == PLUS_INFINITY);
+      assert(realDomain.getSize() == (unsigned int) PLUS_INFINITY);
 
       IntervalIntDomain intDomain(10, 20);
       assertTrue(intDomain.isFinite());
